move unit tests to unit test folder
